Create assignment2.py
With this code we achieve the main goal.The program uses the By.XPATH locator to find the input fields. This locator is based on the XPath expression of the element, which is a unique identifier that will not change if the position or other properties of the element change.